[inclusion] found out how to implement inclusion!

+++ app/src/main/java/com/example/user/watereddown/file_aes.java
@@ -42,21 +42,30 @@ public class file_aes {
 
     public void encryptFile(File file){
         File encrypted = new File(file.getPath() + file.getName() +"_encrypted."+FilenameUtils.getExtension(file.getPath()));
+        if(!encrypted.exists()) {
+            try {
+                if (encrypted.createNewFile())
+                    Log.w("file?", "new!");
+            } catch(Exception e){
+                Log.w("error", e.getMessage());
+            }
+        } else {
 //        if(encrypted.canWrite() && encrypted.canRead())
 //            Log.w("encrypt file?", "yes!");
 //        else Log.w("encrypt file?", "no!");
-        Cipher cp = filealgo.getCipher();
-        SecretKeySpec k = filealgo.getKey();
-        try{
-            FileInputStream in = new FileInputStream(encrypted);
-            cp.init(Cipher.ENCRYPT_MODE, k);
-            CipherOutputStream os = new CipherOutputStream(new FileOutputStream(encrypted),
-                    cp);
-            copy(in, os);
-            in.close();
-            os.close();
-        } catch(Exception ex){
-            Log.w("error", ex.toString());
+            Cipher cp = filealgo.getCipher();
+            SecretKeySpec k = filealgo.getKey();
+            try {
+                FileInputStream in = new FileInputStream(encrypted);
+                cp.init(Cipher.ENCRYPT_MODE, k);
+                CipherOutputStream os = new CipherOutputStream(new FileOutputStream(encrypted),
+                        cp);
+                copy(in, os);
+                in.close();
+                os.close();
+            } catch (Exception ex) {
+                Log.w("error", ex.toString());
+            }
         }
     }
